+++ test/pattern_test_test.rb
@@ -64,6 +64,25 @@ class PatternTestTest < Minitest::Test
     end
   end
 
+  def test_string
+    nodes = E::Literal.new(type: :string, values: ["foo"])
+    assert nodes.test_node(ruby('"foo"'))
+    refute nodes.test_node(ruby('"bar"'))
+  end
+
+  def test_string2
+    nodes = E::Literal.new(type: :string, values: ["foo", "bar"])
+    assert nodes.test_node(ruby('"foo"'))
+    assert nodes.test_node(ruby('"bar"'))
+    refute nodes.test_node(ruby('"baz"'))
+  end
+
+  def test_string3
+    nodes = E::Literal.new(type: :string, values: [/foo/])
+    assert nodes.test_node(ruby('"foo bar"'))
+    refute nodes.test_node(ruby('"baz"'))
+  end
+
   def test_call_without_args
     nodes = query_pattern("foo", "foo(); foo(1)")
     assert_equal 2, nodes.size
@@ -285,5 +304,10 @@ class PatternTestTest < Minitest::Test
   def test_self
     nodes = query_pattern("self.f", "f(); self.f(); foo.f()")
     assert_equal Set.new([ruby("self.f"), ruby("f()")]), Set.new(nodes)
+  end
+
+  def test_string_value
+    nodes = query_pattern("has_many(:symbol: as 'children)", "has_many(:children); has_many(:repositories)", where: { children: [:children] })
+    assert_equal Set.new([ruby("has_many :children")]), Set.new(nodes)
   end
 end