+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master
 
 * Add Wiki pages to repository in manual directory #25
+* Add named literal pattern `:string: as 'name` with `where: { name: ["alice", /bob/] }` #24
 
 ## 0.6.0 (2017-06-27)